[win32] adjust output directory of .obj files
This is necessary to avoid overwriting .obj files of classes with the same name
(independent of the namespace). VS puts all the .obj files into the same place
($(IntDir) by default). By changing that to $(IntDir)\%(RelativeDir) every
.obj file is placed into a relative directory dependong on the hierarchy of
the VS project. The "\a\a" in there are necessary because our VS project files
are placed two levels above the base dir so VS adds "\..\.." in the path
generated by %(RelativeDir).