[fix] (hack) to make sure the python parser doesn't call the AnnouncementManager until it's initialized.
This is a hack to delay the use of the Announcer until after the constructor is called. The XBPython compilation unit seems to now be initialized later than the Announcement code and therefore the static call to CAnnounementManager::AddAnnouncer originally called from the constructor of XBPython (which is called when the XBPython.o compilation unit initializes due to the global g_pythonParser) is executed prior to the static elements of AnnouncementManager.o (one of which is a std::vector) have been constructed. The should be fixed using Dependency Injection.